.history/requirements_20250927172618.txt
# FastAPI Core
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.2
pydantic-settings==2.1.0
python-multipart==0.0.6

# Supabase Database (ONLY DATABASE CONNECTION)
supabase==2.3.4
python-dotenv==1.0.0

# Security & Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Geo Processing
geopy==2.4.0
shapely==2.0.2

# AI/ML Libraries
scikit-learn==1.3.2
pandas==2.1.4
numpy==1.25.2

# Testing & Development
faker==22.0.0
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2